trivial change remove redundant cast.
svn path=/dasCore/trunk/; revision=4154

+++ dasCore/src/edu/uiowa/physics/pw/das/graph/SpectrogramRenderer.java
@@ -282,7 +282,7 @@ public class SpectrogramRenderer extends Renderer implements TableDataSetConsume
                     } else {
                         index = (i - 0) + (ny - j - 1) * nx;
                     }
-                    icolor = (int) cb.indexColorTransform(rebinData.getDouble(i, j, units), units);
+                    icolor = cb.indexColorTransform(rebinData.getDouble(i, j, units), units);
                     pix[index] = (byte) icolor;
                 }
             }
